requirements.txt
Django==1.9.2
djangorestframework==3.3.2
mysqlclient==1.3.7
PyMySQL3==0.5
wheel==0.24.0
gunicorn==19.4.5

Flask==0.10.1
Flask-Sockets==0.2.0
gevent==1.0.2
gevent-websocket==0.9.5
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
redis==2.10.5
Werkzeug==0.11.3


Flask_SocketIO



python-engineio
python-socketio
six==1.9.0
